requirements.txt
asgiref==3.5.2
astroid==2.5
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.5
Django==4.0.5
djangorestframework==3.13.1
Faker==4.14.2
flake8==3.8.1
iniconfig==1.1.1
isort==4.3.21
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.2
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.1
pytz==2022.1
six==1.15.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
validate-docbr==1.8.2
wrapt==1.12.1
